requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
openai>=1.3.0
snownlp>=0.12.3
lxml>=4.9.0
